New upstream version 91.5.1

+++ js/src/jsapi-tests/testArrayBuffer.cpp
@@ -8,6 +8,8 @@
 #include "js/Array.h"        // JS::NewArrayObject
 #include "js/ArrayBuffer.h"  // JS::{GetArrayBuffer{ByteLength,Data},IsArrayBufferObject,NewArrayBuffer{,WithContents},StealArrayBufferContents}
 #include "js/Exception.h"
+#include "js/experimental/TypedData.h"  // JS_New{Int32,Uint8}ArrayWithBuffer
+#include "js/friend/ErrorMessages.h"    // JSMSG_*
 #include "js/MemoryFunctions.h"
 #include "jsapi-tests/tests.h"
 
@@ -173,7 +175,7 @@ BEGIN_TEST(testArrayBuffer_customFreeFunc) {
   CHECK(buffer);
   CHECK(!data.wasFreed());
 
-  uint32_t len;
+  size_t len;
   bool isShared;
   uint8_t* bufferData;
   JS::GetArrayBufferLengthAndData(buffer, &len, &isShared, &bufferData);
@@ -199,7 +201,7 @@ BEGIN_TEST(testArrayBuffer_staticContents) {
   CHECK(buffer);
   CHECK(!data.wasFreed());
 
-  uint32_t len;
+  size_t len;
   bool isShared;
   uint8_t* bufferData;
   JS::GetArrayBufferLengthAndData(buffer, &len, &isShared, &bufferData);
